Crate xdrgen

XDR codec generation

This crate provides library interfaces for programatically generating Rust code to implement RFC4506 XDR encoding/decoding, as well as a command line tool "xdrgen".

It is intended to be used with the "xdr-codec" crate, which provides the runtime library for encoding/decoding primitive types, strings, opaque data and arrays.

Functions

compile
Simplest possible way to generate Rust code from an XDR specification.

generate
Generate Rust code from an RFC4506 XDR specification